ExcelReporter now works with DataBlocks instead of DataAnalyzerReport
