+++ src/QuizData.Analyser/Models/DataBlocks/DistributionDataBlock.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace QuizData.ExcelReport
+namespace QuizData.Analyser.Models.DataBlocks
 {
     public class DistributionDataBlock<TKey, TValue> : IDataBlock
     {
