+++ src/QuizData.Analyser/Models/DataBlocks/IDataBlock.cs
@@ -1,0 +1,6 @@
+﻿namespace QuizData.Analyser.Models.DataBlocks
+{
+    public interface IDataBlock
+    {
+    }
+}
