--- src/QuizData.ExcelReporter/IDataBlock.cs
@@ -1,6 +1,0 @@
-﻿namespace QuizData.ExcelReport
-{
-    public interface IDataBlock
-    {
-    }
-}
